requirements.txt
aiogram
docxtpl
python-dotenv
pypdf
openpyxl
